solve currency FCFA issue  15-11-2024 by Keshav Savita

+++ app/src/main/java/com/my/afarycode/OnlineShopping/orderdetails/ItemsAdapter.java
@@ -42,7 +42,7 @@ public class ItemsAdapter extends RecyclerView.Adapter<ItemsAdapter.MyViewHolder
      //   holder.binding.tvProductTotal.setText(Integer.parseInt(arrayList.get(position).getQuantity()) * Double.parseDouble(arrayList.get(position).getPrice())+"");
 
         holder.binding.tvProductPrice.setText(arrayList.get(position).getQuantity() + " X " + arrayList.get(position).getLocalPrice() );
-        holder.binding.tvProductTotal.setText(Integer.parseInt(arrayList.get(position).getQuantity()) * Double.parseDouble(arrayList.get(position).getLocalPrice())+"");
+        holder.binding.tvProductTotal.setText(Integer.parseInt(arrayList.get(position).getQuantity()) * Integer.parseInt(arrayList.get(position).getLocalPrice())+"");
 
 
     }
